Rename front-end variables
Some variables used in the front-end had ambiguous names, and so have
been renamed appropriately. A comment has also been included at the
start of the program to explain what it does and how to run it. More
comments have also been written throughout the program to allow for
better understanding of how it works.